+++ packages/web/src/props/Radius/utils/getResponsiveRadii.ts
@@ -1,7 +1,7 @@
 import { RadiusProps } from '@shallot-ui/theme'
+import { Getter } from '@shallot-ui/core'
 
-import { Getter } from '../../../types'
-import { getMediaQueries } from '../../../utils'
+import { getMediaQueries } from '../../../utils/getMediaQueries'
 import getStaticRadius from './getStaticRadius'
 
 const getResponsiveRadii: Getter<RadiusProps> = ({ radius }) =>