+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,12 @@
 # stories
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/web@0.32.0
+
 ## 0.2.2
 
 ### Patch Changes