+++ packages/next/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @shallot-ui/next
 
+## 0.24.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/web@0.32.0
+
 ## 0.24.2
 
 ### Patch Changes