+++ packages/react-select/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @shallot-ui/react-select
 
+## 0.25.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/web@0.32.0
+
 ## 0.25.2
 
 ### Patch Changes